# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DTI O Salnes — API Publica
  description: |
    API REST publica de la Plataforma Digital Turistica Inteligente de la Mancomunidad de O Salnes.
    Modelo de datos alineado con UNE 178503. Interoperabilidad con PID SEGITTUR.
  version: 1.0.0
  contact:
    name: Mancomunidad de O Salnes
    url: https://turismo.osalnes.gal

servers:
  - url: https://oduglbxjcmmdexwplzvw.supabase.co/functions/v1/api
    description: Produccion (Supabase Edge Functions)
  - url: http://localhost:3001/api/v1
    description: Desarrollo local

tags:
  - name: Recursos
    description: Recursos turisticos (UNE 178503)
  - name: Tipologias
    description: Catalogo de tipologias UNE 178503
  - name: Categorias
    description: Categorias jerarquicas de recursos
  - name: Municipios
    description: Municipios de la Mancomunidad
  - name: Zonas
    description: Zonas geograficas dentro de municipios
  - name: Mapa
    description: Endpoints optimizados para el mapa interactivo
  - name: Busqueda
    description: Busqueda de texto en recursos
  - name: Paginas
    description: Paginas de contenido editorial
  - name: Navegacion
    description: Menus de navegacion del portal
  - name: Eventos
    description: Recursos de tipo evento

paths:
  /resources:
    get:
      tags: [Recursos]
      summary: Listar recursos turisticos
      parameters:
        - name: type
          in: query
          schema: { type: string }
          description: Filtrar por tipologia (e.g. Hotel, Beach, Restaurant)
        - name: municipio
          in: query
          schema: { type: string, format: uuid }
          description: Filtrar por municipio ID
        - name: status
          in: query
          schema: { type: string, enum: [borrador, revision, publicado, archivado] }
        - name: lang
          in: query
          schema: { type: string, enum: [es, gl, en, fr, pt] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: sort
          in: query
          schema: { type: string, enum: [created_at, updated] }
      responses:
        '200':
          description: Lista paginada de recursos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/TouristResource' }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
                  pages: { type: integer }

  /resources/{id}:
    get:
      tags: [Recursos]
      summary: Detalle de recurso por ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Recurso completo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TouristResource' }
        '404':
          description: Recurso no encontrado

  /resources/by-slug/{slug}:
    get:
      tags: [Recursos]
      summary: Detalle de recurso por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Recurso completo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TouristResource' }
        '404':
          description: Recurso no encontrado

  /typologies:
    get:
      tags: [Tipologias]
      summary: Catalogo de tipologias UNE 178503
      responses:
        '200':
          description: Lista de tipologias
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Typology' }

  /categories:
    get:
      tags: [Categorias]
      summary: Arbol de categorias
      responses:
        '200':
          description: Lista plana de categorias (parent_id para jerarquia)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }

  /categories/{slug}:
    get:
      tags: [Categorias]
      summary: Detalle de categoria por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Categoria
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404':
          description: Categoria no encontrada

  /municipalities:
    get:
      tags: [Municipios]
      summary: Municipios de O Salnes
      responses:
        '200':
          description: Lista de municipios
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Municipality' }

  /zones:
    get:
      tags: [Zonas]
      summary: Zonas geograficas
      parameters:
        - name: municipio
          in: query
          schema: { type: string, format: uuid }
          description: Filtrar por municipio
      responses:
        '200':
          description: Lista de zonas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Zone' }

  /map/resources:
    get:
      tags: [Mapa]
      summary: Recursos para el mapa interactivo (bounding box)
      description: |
        Devuelve hasta 500 recursos publicados y visibles en mapa dentro del bounding box.
        Incluye grupo de tipologia para iconos coloreados y nombres traducidos.
      parameters:
        - name: bounds
          in: query
          required: true
          schema: { type: string }
          description: "Bounding box: lat_sw,lng_sw,lat_ne,lng_ne"
          example: "42.3,-8.9,42.6,-8.6"
        - name: type
          in: query
          schema: { type: string }
          description: Filtrar por tipologia
        - name: municipio
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Recursos dentro del bounding box
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MapResource' }
        '400':
          description: Parametro bounds invalido

  /search:
    get:
      tags: [Busqueda]
      summary: Buscar recursos por texto
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2 }
          description: Texto a buscar (minimo 2 caracteres)
        - name: lang
          in: query
          schema: { type: string, enum: [es, gl, en, fr, pt] }
        - name: type
          in: query
          schema: { type: string }
        - name: municipio
          in: query
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Resultados de busqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/SearchResult' } }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
                  pages: { type: integer }

  /pages/{slug}:
    get:
      tags: [Paginas]
      summary: Pagina de contenido por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pagina con contenido traducido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
        '404':
          description: Pagina no encontrada

  /navigation/{menuSlug}:
    get:
      tags: [Navegacion]
      summary: Menu de navegacion por slug
      parameters:
        - name: menuSlug
          in: path
          required: true
          schema: { type: string }
          example: header
      responses:
        '200':
          description: Items del menu
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NavItem' }

  /events:
    get:
      tags: [Eventos]
      summary: Listado de eventos
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    slug: { type: string }
                    tipologia: { type: object }
                    latitude: { type: number, nullable: true }
                    longitude: { type: number, nullable: true }

components:
  schemas:
    LocalizedValue:
      type: object
      description: Valor multilingue (claves es, gl, en, fr, pt)
      additionalProperties: { type: string }
      example: { es: "Hotel Cambados", gl: "Hotel Cambados" }

    TouristResource:
      type: object
      description: Recurso turistico segun UNE 178503
      properties:
        id: { type: string, format: uuid }
        uri: { type: string, description: "URI semantica (osalnes:recurso:slug)" }
        rdfType: { type: string, description: "Tipologia UNE 178503 (Hotel, Beach, etc.)" }
        slug: { type: string }
        name: { $ref: '#/components/schemas/LocalizedValue' }
        description: { $ref: '#/components/schemas/LocalizedValue' }
        seoTitle: { $ref: '#/components/schemas/LocalizedValue' }
        seoDescription: { $ref: '#/components/schemas/LocalizedValue' }
        location:
          type: object
          properties:
            latitude: { type: number, nullable: true }
            longitude: { type: number, nullable: true }
            streetAddress: { type: string, nullable: true }
            postalCode: { type: string, nullable: true }
        municipioId: { type: string, format: uuid, nullable: true }
        zonaId: { type: string, format: uuid, nullable: true }
        contact:
          type: object
          properties:
            telephone: { type: array, items: { type: string } }
            email: { type: array, items: { type: string } }
            url: { type: string, nullable: true }
            sameAs: { type: array, items: { type: string } }
        touristTypes: { type: array, items: { type: string }, description: "UNE 178503 sec. 7.6" }
        ratingValue: { type: integer, minimum: 1, maximum: 6, nullable: true }
        servesCuisine: { type: array, items: { type: string } }
        isAccessibleForFree: { type: boolean, nullable: true }
        publicAccess: { type: boolean, nullable: true }
        occupancy: { type: integer, nullable: true }
        openingHours: { type: string, nullable: true }
        extras: { type: object, description: "UNE 178503 sec. 7.8 datos extendidos" }
        status: { type: string, enum: [borrador, revision, publicado, archivado] }
        visibleOnMap: { type: boolean }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        categoryIds: { type: array, items: { type: string, format: uuid } }

    MapResource:
      type: object
      description: Recurso simplificado para el mapa
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { $ref: '#/components/schemas/LocalizedValue' }
        rdfType: { type: string }
        grupo: { type: string, nullable: true, description: "Grupo de tipologia (alojamiento, restauracion, recurso, evento, transporte, servicio)" }
        location:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
            streetAddress: { type: string, nullable: true }

    SearchResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        uri: { type: string }
        slug: { type: string }
        tipologia: { type: object }
        municipioId: { type: string, format: uuid, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        name: { $ref: '#/components/schemas/LocalizedValue' }

    Typology:
      type: object
      description: Tipologia UNE 178503 sec. 7.5
      properties:
        id: { type: string, format: uuid }
        typeCode: { type: string, description: "Codigo (Hotel, Beach, Restaurant, etc.)" }
        schemaOrgType: { type: string, nullable: true }
        grupo: { type: string, description: "alojamiento, restauracion, recurso, evento, transporte, servicio" }
        name: { $ref: '#/components/schemas/LocalizedValue' }

    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        parentId: { type: string, format: uuid, nullable: true }
        orden: { type: integer }
        name: { $ref: '#/components/schemas/LocalizedValue' }

    Municipality:
      type: object
      properties:
        id: { type: string, format: uuid }
        codigoIne: { type: string, description: "Codigo INE del municipio" }
        slug: { type: string }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        name: { $ref: '#/components/schemas/LocalizedValue' }

    Zone:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        municipioId: { type: string, format: uuid }
        name: { $ref: '#/components/schemas/LocalizedValue' }

    Page:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        template: { type: string, nullable: true }
        title: { $ref: '#/components/schemas/LocalizedValue' }
        body: { $ref: '#/components/schemas/LocalizedValue' }
        seoTitle: { $ref: '#/components/schemas/LocalizedValue' }
        seoDescription: { $ref: '#/components/schemas/LocalizedValue' }

    NavItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        tipo: { type: string }
        referencia: { type: string }
        orden: { type: integer }
        label: { $ref: '#/components/schemas/LocalizedValue' }